feat: migration react router query
