--- web/.react-router/types/+register.ts
@@ -1,12 +1,0 @@
-import "react-router";
-
-declare module "react-router" {
-  interface Register {
-    params: Params;
-  }
-}
-
-type Params = {
-  "/": {};
-  "/register": {};
-};
