--- web/app/lib/types.ts
@@ -1,44 +1,0 @@
-// Types pour CVMaker
-export interface PersonalInfo {
-  firstName: string;
-  lastName: string;
-  email: string;
-  phone: string;
-  address: string;
-  linkedIn?: string;
-  website?: string;
-}
-
-export interface Experience {
-  id: string;
-  company: string;
-  position: string;
-  startDate: string;
-  endDate?: string;
-  isCurrentJob: boolean;
-  description: string;
-}
-
-export interface Education {
-  id: string;
-  institution: string;
-  degree: string;
-  field: string;
-  startDate: string;
-  endDate?: string;
-  isCurrentStudy: boolean;
-}
-
-export interface Skill {
-  id: string;
-  name: string;
-  level: 'beginner' | 'intermediate' | 'advanced' | 'expert';
-}
-
-export interface CVData {
-  personalInfo: PersonalInfo;
-  experiences: Experience[];
-  education: Education[];
-  skills: Skill[];
-  summary?: string;
-}
