--- web/app/routes.ts
@@ -1,6 +1,0 @@
-import { type RouteConfig, index, route } from "@react-router/dev/routes";
-
-export default [
-  index("routes/home.tsx"),
-  route("/register", "./routes/register.tsx"),
-] satisfies RouteConfig;
