+++ web/package.json
@@ -3,10 +3,10 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "react-router build",
-    "dev": "react-router dev",
-    "start": "react-router-serve ./build/server/index.js",
-    "typecheck": "react-router typegen && tsc"
+    "build": "vite build",
+    "dev": "vite",
+    "preview": "vite preview",
+    "typecheck": "tsc"
   },
   "dependencies": {
     "@hookform/resolvers": "^5.1.1",
@@ -14,17 +14,16 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.60.0",
-    "react-router": "^7.5.1",
+    "react-router-dom": "^6.30.1",
     "zod": "^4.0.2",
     "zustand": "^5.0.6"
   },
   "devDependencies": {
-    "@react-router/dev": "^7.5.1",
     "@tailwindcss/vite": "^4.0.0",
     "@types/node": "^20",
     "@types/react": "^19.0.1",
     "@types/react-dom": "^19.0.1",
-    "react-router-devtools": "^1.1.0",
+    "@vitejs/plugin-react": "^4.6.0",
     "tailwindcss": "^4.0.0",
     "typescript": "^5.7.2",
     "vite": "^5.4.11",
