+++ web/src/routes/home.tsx
@@ -1,0 +1,8 @@
+export default function Home() {
+  return (
+    <div>
+      <h1>CV MAKER MOFO</h1>
+      <p>Parce que chercher un job ça casse les couilles</p>
+    </div>
+  );
+}
